requirements.txt
asgiref==3.8.1
Django==5.0.1
django-admin-thumbnails==0.2.9
django-adminlte-3==0.1.6
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-mptt==0.16.0
pillow==10.4.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
whitenoise==6.7.0
